% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{iri_company}
\alias{iri_company}
\title{iri_company}
\usage{
iri_company(company)
}
\arguments{
\item{company}{The name of a company.}
}
\value{
Company's name.
}
\description{
This function allows you to find and search the right company name associated with the Industrial Research and Innovation's Data.
If no argument is filed, all names will be displayed.
}
\examples{
mycompany<- iri_company()
mycompany<- iri_company(company = "Samsung")
mycompany<- iri_company("Samsung")

}
\seealso{
\code{\link{iri_country}} for the IRI's country code, \code{\link{iri_indicator}} for the IRI's indicators, \code{\link{iri_industry}} for the IRI's industries name and \code{\link{iri_data}} to collect the data.
}
