% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{iri_country}
\alias{iri_country}
\title{iri_country}
\usage{
iri_country(country)
}
\arguments{
\item{country}{The name of the country.}
}
\value{
Country's ISO code.
}
\description{
This function allows you to find and search the right country code associated with the Industrial Research and Innovation's Data.
If no argument is filed, all indicators will be displayed.
}
\examples{
mycountry <- iri_country()
mycountry <- iri_country(country = "Canada")
mycountry <- iri_country("Canada")

}
\seealso{
\code{\link{iri_indicator}} for the IRI's indicators, \code{\link{iri_company}} for the IRI's companies name, \code{\link{iri_industry}} for the IRI's industries name and \code{\link{iri_data}} to collect the data.
}
