% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{iri_indicator}
\alias{iri_indicator}
\title{iri_indicator}
\usage{
iri_indicator(indicators)
}
\arguments{
\item{indicators}{An indicator from the Industrial Research and Innovation.}
}
\value{
Indicator code from the Industrial Research and Innovation.
}
\description{
This function allows you to find and search the right indicator code from the Industrial Research and Innovation you want to use.
If no argument is filed, all indicators will be displayed.
}
\examples{
myIndicator <- iri_indicator()
myIndicator <- iri_indicator(indicators = "sales")
myIndicator <- iri_indicator("sales")

}
\seealso{
\code{\link{iri_country}} for the IRI's country code, \code{\link{iri_company}} for the IRI's companies name, \code{\link{iri_industry}} for the IRI's industries name and \code{\link{iri_data}} to collect the data.
}
