% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{iri_industry}
\alias{iri_industry}
\title{iri_industry}
\usage{
iri_industry(industry)
}
\arguments{
\item{industry}{The name of the industrial sector}
}
\value{
Industry's name.
}
\description{
This function allows you to find and search the right industry name associated with the Industrial Research and Innovation's Data.
If no argument is filed, all names will be displayed.
}
\examples{
myindustry <- iri_industry()
myindustry <- iri_industry(industry = "Automobile")
myindustry <- iri_industry("Automobile")

}
\seealso{
\code{\link{iri_country}} for the IRI's country code, \code{\link{iri_indicator}} for the IRI's indicators, \code{\link{iri_company}} for the IRI's companies name and \code{\link{iri_data}} to collect the data.
}
