% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{irir_company}
\alias{irir_company}
\title{irir_company}
\usage{
irir_company(company)
}
\arguments{
\item{company}{The name of a company.}
}
\value{
Company's name.
}
\description{
This function allows you to find and search the right company name associated with the Industrial Research and Innovation's Data.
If no argument is filed, all names will be displayed.
}
\examples{
mycompany<- irir_company(company = "Samsung")
}
\seealso{
\code{\link{irir_country}} for the IRI's country code, \code{\link{irir_indicator}} for the IRI's indicators, \code{\link{irir_industry}} for the IRI's industries name and \code{\link{irir_data}} to collect the data.
}
