% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{irir_country}
\alias{irir_country}
\title{irir_country}
\usage{
irir_country(country)
}
\arguments{
\item{country}{The name of the country.}
}
\value{
Country's ISO code.
}
\description{
This function allows you to find and search the right country code associated with the Industrial Research and Innovation's Data.
If no argument is filed, all indicators will be displayed.
}
\examples{
mycountry <- irir_country(country = "Canada")
}
\seealso{
\code{\link{irir_indicator}} for the IRI's indicators, \code{\link{irir_company}} for the IRI's companies name, \code{\link{irir_industry}} for the IRI's industries name and \code{\link{irir_data}} to collect the data.
}
