% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{irir_indicator}
\alias{irir_indicator}
\title{irir_indicator}
\usage{
irir_indicator(indicators)
}
\arguments{
\item{indicators}{An indicator from the Industrial Research and Innovation.}
}
\value{
Indicator code from the Industrial Research and Innovation.
}
\description{
This function allows you to find and search the right indicator code from the Industrial R&D Investment Scoreboard you want to use.
If no argument is filed, all indicators will be displayed.
}
\examples{
myIndicator <- irir_indicator(indicators = "sales")
}
\seealso{
\code{\link{irir_country}} for the IRI's country code, \code{\link{irir_company}} for the IRI's companies name, \code{\link{irir_industry}} for the IRI's industries name and \code{\link{irir_data}} to collect the data.
}
