% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{irir_industry}
\alias{irir_industry}
\title{irir_industry}
\usage{
irir_industry(industry)
}
\arguments{
\item{industry}{The name of the industrial sector.}
}
\value{
Industry's name.
}
\description{
This function allows you to find and search the right industry name associated with the Industrial Research and Innovation's Data.
If no argument is filed, all names will be displayed.
}
\examples{
myindustry <- irir_industry(industry = "Automobile")
}
\seealso{
\code{\link{irir_country}} for the IRI's country code, \code{\link{irir_indicator}} for the IRI's indicators, \code{\link{irir_company}} for the IRI's companies name and \code{\link{irir_data}} to collect the data.
}
