% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{irir_visual}
\alias{irir_visual}
\title{irir_visual}
\usage{
irir_visual(
  country = "CAN",
  chart = "bar_1",
  title = TRUE,
  years = as.numeric(max(IRI_data$year))
)
}
\arguments{
\item{country}{The Country ISO code}

\item{chart}{Type of charts}

\item{title}{Chart title, set by default to TRUE}

\item{years}{Year, only works for bar chart and set by default to max year of IRI's data}
}
\value{
Chosen Graph
}
\description{
This function allows you to create 3 sorts of visuals: line, bar and point charts.
}
\examples{
irir_visual(country = "CAN", chart = "bar_1", title = TRUE, years = 2019)
}
\seealso{
\code{\link{irir_country}} for the IRI's country code, \code{\link{irir_industry}} for the IRI's industries name, \code{\link{irir_indicator}} for the IRI's indicators, \code{\link{irir_company}} for the IRI's companies name and \code{\link{irir_data}} to collect the data.
}
