% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{sqs_iri_company}
\alias{sqs_iri_company}
\title{sqs_iri_company}
\usage{
sqs_iri_company(company)
}
\arguments{
\item{company}{The name of a company.}
}
\value{
Company's name.
}
\description{
This function allows you to find and search the right company name associated with the Industrial Research and Innovation's Data.
If no argument is filed, all names will be displayed.
}
\examples{
mycompany<- sqs_iri_company()
mycompany<- sqs_iri_company(company = "Samsung")
mycompany<- sqs_iri_company("Samsung")

}
\seealso{
\code{\link{sqs_iri_country}} for the IRI's country code, \code{\link{sqs_iri_indicator}} for the IRI's indicators, \code{\link{sqs_iri_industry}} for the IRI's industries name and \code{\link{sqs_iri_data}} to collect the data.
}
