% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{sqs_iri_country}
\alias{sqs_iri_country}
\title{sqs_iri_country}
\usage{
sqs_iri_country(country)
}
\arguments{
\item{country}{The name of the country.}
}
\value{
Country's ISO code.
}
\description{
This function allows you to find and search the right country code associated with the Industrial Research and Innovation's Data.
If no argument is filed, all indicators will be displayed.
}
\examples{
mycountry <- sqs_iri_country()
mycountry <- sqs_iri_country(country = "Canada")
mycountry <- sqs_iri_country("Canada")

}
\seealso{
\code{\link{sqs_iri_indicator}} for the IRI's indicators, \code{\link{sqs_iri_company}} for the IRI's companies name, \code{\link{sqs_iri_industry}} for the IRI's industries name and \code{\link{sqs_iri_data}} to collect the data.
}
