% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{sqs_iri_data}
\alias{sqs_iri_data}
\title{sqs_iri_data}
\usage{
sqs_iri_data(country = data_long_country, years = data_long_year,
  indicators = dat_long_indicator, company = data_long_company,
  industry = data_long_industry, rank = data_long_rank)
}
\arguments{
\item{country}{Countries' ISO code.}

\item{years}{Years for which you want the data.}

\item{indicators}{Indicators from the Industrial Research and Innovation.}

\item{company}{Companies for which you want the data.}

\item{industry}{Industries for which you want the data.}

\item{rank}{Rank of a company.}
}
\value{
Data for the country, indicator, year, company, industrial sector and rank requested
}
\description{
This function allows you to find and display the Industrial Research and Innovation data according to the selected parameters.
If no arguments are filled, all data will be displayed.
}
\examples{
myData <- sqs_iri_data(country = "USA", years = "2018", indicators = "RD.euro", company = "FORD MOTOR")
myData <- sqs_iri_data("USA", "2018", "RD.euro", "FORD MOTOR", "Automobiles & Parts", "14")
myData <- sqs_iri_data(country =c("USA","DEU"), years =c("2018"), rank = 1:25 )

}
\seealso{
\code{\link{sqs_iri_indicator}} for the IRI's indicator symbol, \code{\link{sqs_iri_country}} for the country's ISO code, \code{\link{sqs_iri_company}} for the IRI's companies name and \code{\link{sqs_iri_industry}} for the IRI's industries name.
}
