% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iriR.R
\name{sqs_iri_industry}
\alias{sqs_iri_industry}
\title{sqs_iri_industry}
\usage{
sqs_iri_industry(industry)
}
\arguments{
\item{industry}{The name of the industrial sector}
}
\value{
Industry's name.
}
\description{
This function allows you to find and search the right industry name associated with the Industrial Research and Innovation's Data.
If no argument is filed, all names will be displayed.
}
\examples{
myindustry <- sqs_iri_industry()
myindustry <- sqs_iri_industry(industry = "Automobile")
myindustry <- sqs_iri_industry("Automobile")

}
\seealso{
\code{\link{sqs_iri_country}} for the IRI's country code, \code{\link{sqs_iri_indicator}} for the IRI's indicators, \code{\link{sqs_iri_company}} for the IRI's companies name and \code{\link{sqs_iri_data}} to collect the data.
}
